Move Container implementations to ContainerImpls.(h|c)pp
