+++ CMakeLists.txt
@@ -415,8 +415,8 @@ set(CORE_SOURCE
         src/backend/Writable.cpp
         src/benchmark/mpi/OneDimensionalBlockSlicer.cpp
         src/helper/list_series.cpp
+        src/snapshots/ContainerImpls.cpp
         src/snapshots/IteratorTraits.cpp
-        src/snapshots/Snapshots.cpp
         src/snapshots/StatefulIterator.cpp)
 set(IO_SOURCE
         src/IO/AbstractIOHandler.cpp
