+++ src/snapshots/ContainerImpls.cpp
@@ -1,14 +1,9 @@
-#include "openPMD/snapshots/Snapshots.hpp"
+#include "openPMD/snapshots/ContainerImpls.hpp"
 #include "openPMD/snapshots/StatefulIterator.hpp"
 #include <stdexcept>
 
 namespace openPMD
 {
-// OpaqueSeriesIterator Snapshots::begin()
-// {
-//     // return OpaqueSeriesIterator();
-//     throw std::runtime_error("unimplemented");
-// }
 auto StatefulSnapshotsContainer::begin() -> iterator_t
 {
     return m_begin();
